Hard code CUDA to use uint8_t for computing depth

+++ include/cuda_proj.hpp
@@ -4,10 +4,12 @@
 
 #include <cuda_runtime.h>
 
+#include <stdint.h>
+
 namespace octomap_depth_mapping
 {
 
-void project_depth_img(ushort*, double*, int, int,
+void project_depth_img(uint8_t*, double*, int, int,
     dim3, dim3,
     double, double, double, double,
     double, double, double,
@@ -15,7 +17,7 @@ void project_depth_img(ushort*, double*, int, int,
     double, double, double,
     double, double, double);
 
-__global__ void project_kernel(ushort*, double*, int, int,
+__global__ void project_kernel(uint8_t*, double*, int, int,
     double, double, double, double,
     double, double, double,
     double, double, double,
